feat: load more data in agenda
