+++ .env.test
@@ -1,1 +1,2 @@
 SUPPORT_MAIL=support@pass-emploi.beta.gouv.fr
+IDS_STRUCTURES_EARLY_ADOPTERS=id-test
