+++ components/rdv/TableauRdvsConseiller.tsx
@@ -1,6 +1,8 @@
-import React from 'react'
+import { DateTime } from 'luxon'
+import React, { useState } from 'react'
 
 import { RdvRow } from 'components/rdv/RdvRow'
+import Button, { ButtonStyle } from 'components/ui/Button/Button'
 import Table from 'components/ui/Table/Table'
 import { TBody } from 'components/ui/Table/TBody'
 import { TH } from 'components/ui/Table/TH'
@@ -8,15 +10,18 @@ import { THead } from 'components/ui/Table/THead'
 import { TR } from 'components/ui/Table/TR'
 import { EvenementListItem } from 'interfaces/evenement'
 import { AgendaData, renderAgenda } from 'presentation/Intercalaires'
+import { toFrenchFormat, WEEKDAY_MONTH_LONG } from 'utils/date'
 
 type TableauRdvsConseillerProps = {
   idConseiller: string
   agendaRdvs: AgendaData<EvenementListItem>
+  onChargerEvenementsJour: (jour: DateTime) => Promise<void>
 }
 
 export default function TableauRdvsConseiller({
   agendaRdvs,
   idConseiller,
+  onChargerEvenementsJour,
 }: TableauRdvsConseillerProps) {
   return (
     <Table asDiv={true} caption={{ text: 'Liste de mes événements' }}>
@@ -36,9 +41,41 @@ export default function TableauRdvsConseiller({
           (rdv) => (
             <RdvRow key={rdv.id} rdv={rdv} idConseiller={idConseiller} />
           ),
-          true
+          (jourISO: string) => (
+            <ButtonChargerEvenementsJour
+              jour={DateTime.fromISO(jourISO)}
+              onClick={onChargerEvenementsJour}
+            />
+          )
         )}
       </TBody>
     </Table>
+  )
+}
+
+function ButtonChargerEvenementsJour({
+  jour,
+  onClick,
+}: {
+  jour: DateTime
+  onClick: (jour: DateTime) => void
+}): React.JSX.Element {
+  const [isLoading, setIsLoading] = useState<boolean>(false)
+
+  function propagateClick() {
+    setIsLoading(true)
+    onClick(jour)
+  }
+
+  return (
+    <Button
+      style={ButtonStyle.SECONDARY}
+      className='m-auto'
+      onClick={propagateClick}
+      isLoading={isLoading}
+    >
+      Afficher l’agenda du&nbsp;
+      <span aria-label={toFrenchFormat(jour, WEEKDAY_MONTH_LONG)}>jour</span>
+    </Button>
   )
 }
