+++ tests/components/TableauRdvsConseiller.test.tsx
@@ -1,5 +1,6 @@
 import { screen } from '@testing-library/dom'
 import { render, within } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
 import { DateTime } from 'luxon'
 import { useRouter } from 'next/router'
 import React from 'react'
@@ -11,11 +12,11 @@ import { buildAgenda } from 'presentation/Intercalaires'
 import {
   TIME_24_H_SEPARATOR,
   toFrenchFormat,
-  toShortDate,
   WEEKDAY_MONTH_LONG,
 } from 'utils/date'
 
 describe('<TableauRdvsConseiller>', () => {
+  const chargerEvenementsJour: (jour: DateTime) => Promise<void> = jest.fn()
   beforeEach(async () => {
     ;(useRouter as jest.Mock).mockReturnValue({ asPath: '/mes-jeunes' })
   })
@@ -31,11 +32,18 @@ describe('<TableauRdvsConseiller>', () => {
           debut: DateTime.fromISO('2021-10-21'),
           fin: DateTime.fromISO('2021-10-26'),
         },
-        ({ date }) => DateTime.fromISO(date)
+        ({ date }) => DateTime.fromISO(date),
+        [0, 1, 4]
       )
 
       // When
-      render(<TableauRdvsConseiller agendaRdvs={agenda} idConseiller='1' />)
+      render(
+        <TableauRdvsConseiller
+          agendaRdvs={agenda}
+          idConseiller='1'
+          onChargerEvenementsJour={chargerEvenementsJour}
+        />
+      )
     })
 
     it('affiche les informations des rendez-vous', () => {
@@ -86,6 +94,20 @@ describe('<TableauRdvsConseiller>', () => {
       expect(within(rendezVous1).getByText('non')).toHaveAttribute(
         'class',
         'sr-only'
+      )
+    })
+
+    it('permet de charger plus de rendez-vous', async () => {
+      // When
+      await userEvent.click(
+        screen.getByRole('button', {
+          name: 'Afficher l’agenda du samedi 23 octobre',
+        })
+      )
+
+      // Then
+      expect(chargerEvenementsJour).toHaveBeenCalledWith(
+        DateTime.fromISO('2021-10-23')
       )
     })
   })
